+++ README.md
@@ -1,0 +1,20 @@
+# Combine Strings Lab
+
+## Pre-Requisites
+- You are expected to review the `String Basics` written lecture prior to working on this lab. If you are not familiar with the concepts being explored in the lab, please make sure to review the study material first.
+
+## Description
+In this exercise, you will practice concatenating, or combinging, string values.
+
+## Completing the Lab
+You will be editing the file `StringUtilities.java`. This file can be found under the `src/main/java` package. Review the file's code and code comments for further instructions.
+
+Tasks to complete within the `StringUtilities.java` file:
+- [ ] Revise the return statement of the method named `combine` so that it returns the correct string value.
+
+## Running Tests
+If you look within the `src/test/java` folder, you will find the Java class `StringUtilitiesTest.java`. The `src/test/java` folder will contain your test files.  
+
+To run the tests, click on the Testing tab (test beaker icon on the left-hand side) and run the test cases by using the "Run Test" button (play icon). Please ensure that you do not already have the program running.
+
+You can optionally run the program for manual testing by switching to the file named `Main.java` under the `src/main/java` package and clicking the "Run Java" button (play icon) on the top-right of VSCode.
