--- src/main/java/Lab.java
@@ -1,14 +1,0 @@
-public class Lab {
-    /**
-     * TODO: return the combination of two Strings.
-     * You can combine two strings using the '+' operator.
-     * For instance, String bigstr = str1 + str2.
-     * This method should return the combination of two Strings.
-     * @param a the first part of the string to be combined.
-     * @param b the second part of the string to be combined.
-     * @return the combination of strings a and b.
-     */
-    public String combine(String a, String b){
-        return "";
-    }
-}
