+++ src/main/java/StringUtilities.java
@@ -1,0 +1,15 @@
+public class StringUtilities {
+
+    /**
+     * This method returns the concatenated value of its two string parameters.
+     * 
+     * Replace the `null` value with the appropriate expression.
+     * 
+     * @param a - the first part of the string to be combined.
+     * @param b - the second part of the string to be combined.
+     * @return the combination of strings a and b.
+     */
+    public String combine(String a, String b){
+        return null;
+    }
+}
